Abpassungen bei den Berechungen Konzentration und Anreicherung

+++ Weinrechnerlel/Weinrechnerlel/Weinrechnerlel/Models/AnreichRestResponse.cs
@@ -13,8 +13,8 @@ namespace Weinrechnerlel.Models
         public String pa_e { get; set; }
         public double sach { get; set; }
         public double sach_ges { get; set; }
-        public double mehr_an { get; set; }
-        public double menge_an { get; set; }
+        public int mehr_an { get; set; }
+        public int menge_an { get; set; }
         public int EventStatus { get; set; }
         public string EventMessage { get; set; }
