+++ Weinrechnerlel/Weinrechnerlel/Weinrechnerlel/Models/Ergebnis_anreich.cs
@@ -12,7 +12,7 @@ namespace Weinrechnerlel.Models
         public String pa_e { get; set; }
         public double sach { get; set; }
         public double sach_ges { get; set; }
-        public double mehr_an { get; set; }
-        public double menge_an { get; set; }
+        public int mehr_an { get; set; }
+        public int menge_an { get; set; }
     }
 }
